python_example_images/requirements.txt
flask==2.0.1
requests==2.28.2
python-dotenv==1.0.0
